绘制：添加点击双键可以变换双键的方向功能

+++ ChemMaster/ChemMaster/Classes/Tool/CYLBadgeView.m
@@ -43,7 +43,7 @@
 {
     _badgeValue = badgeValue;
     
-    [self setTitle:[NSString stringWithFormat:@"%ld", badgeValue] forState:UIControlStateNormal];
+    [self setTitle:[NSString stringWithFormat:@"%ld", (long)badgeValue] forState:UIControlStateNormal];
 }
 
 @end
